requirements.txt
# Flask e dependências web
Flask==3.0.0
Werkzeug==3.0.1

# Processamento de documentos
python-docx==1.1.0
pdfplumber==0.10.3
reportlab==4.0.7

# IA e APIs
google-genai

# Firebase
firebase-admin==6.4.0

# Utilitários
requests==2.31.0
python-dotenv==1.0.0

# Servidor WSGI para produção
gunicorn==21.2.0

# Dependências de desenvolvimento (opcional)
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0
weasyprint